Make all Turn related types explicit
