+++ src/test-factories/turn.ts
@@ -1,5 +1,11 @@
 import { useRules } from '@/composables/useRules';
-import { useTurnsStore } from '@/stores/turns';
+import {
+	useTurnsStore,
+	type PlayedTurn,
+	type SkippedTurn,
+	type TurnInput,
+	type Turn
+} from '@/stores/turns';
 import { generateId } from '@/utilities/id';
 import { assertActivePinia } from './assertActivePinia';
